qa-test-project-main/requirements.txt
django==4.2
djangorestframework==3.14
pytest==7.4
pytest-django==4.5
model-bakery==1.15
